# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farmacas",
        "version": "0.1"
    },
    "paths": {
...
        "/api/auth/schedules": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Recupera dados de agendamento",
                "description": "Este endpoint permite recuperar dados de agendamento com base em filtros opcionais.",
                "operationId": "f41385d00ff6915c5dfbd274a8af0aff",
                "parameters": [
                    {
                        "name": "starts_at",
                        "in": "query",
                        "description": "Data de início para filtrar os agendamentos (opcional).",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "ends_at",
                        "in": "query",
                        "description": "Data de término para filtrar os agendamentos (opcional).",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Título para filtrar os agendamentos (opcional).",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados de agendamento recuperados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado. O usuário não está autenticado."
                    },
                    "404": {
                        "description": "Usuário não encontrado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Cria um novo agendamento",
                "description": "Este endpoint permite criar um novo agendamento com os campos especificados.",
                "operationId": "c786c50f8f6b7413786857d8654f3f66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Título do agendamento",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição do agendamento",
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "description": "ID do usuário associado ao agendamento",
                                        "type": "integer"
                                    },
                                    "starts_at": {
                                        "description": "Data e hora de início do agendamento (Formato: YYYY-MM-DD HH:MM:SS)",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "ends_at": {
                                        "description": "Data e hora de término do agendamento (Formato: YYYY-MM-DD HH:MM:SS) - Opcional",
                                        "type": "string",
                                        "format": "date-time",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Status do agendamento",
                                        "type": "string"
                                    },
                                    "created_at": {
                                        "description": "Data e hora de criação do agendamento (Formato: YYYY-MM-DD HH:MM:SS)",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_at": {
                                        "description": "Data e hora de atualização do agendamento (Formato: YYYY-MM-DD HH:MM:SS)",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "id": {
                                        "description": "ID do agendamento criado",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agendamento criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "description": "Título do agendamento",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Descrição do agendamento",
                                            "type": "string"
                                        },
                                        "user_id": {
                                            "description": "ID do usuário associado ao agendamento",
                                            "type": "integer"
                                        },
                                        "starts_at": {
                                            "description": "Data e hora de início do agendamento (Formato: YYYY-MM-DD HH:MM:SS)",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "ends_at": {
                                            "description": "Data e hora de término do agendamento (Formato: YYYY-MM-DD HH:MM:SS) - Opcional",
                                            "type": "string",
                                            "format": "date-time",
                                            "nullable": true
                                        },
                                        "status": {
                                            "description": "Status do agendamento",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Data e hora de criação do agendamento (Formato: YYYY-MM-DD HH:MM:SS)",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Data e hora de atualização do agendamento (Formato: YYYY-MM-DD HH:MM:SS)",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "id": {
                                            "description": "ID do agendamento criado",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação. Verifique os campos fornecidos."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}